Restructure code to multiple modules, for adding code in the future
