+++ lib/Transforms/IPO/StructFieldAccessInfo.cpp
@@ -1,0 +1,318 @@
+// lib/Tranforms/IPO/StructFieldAccessInfo.cpp - Implements class of StructFieldAccessInfo -*- C++ -*-===//
+//
+//                     The LLVM Compiler Infrastructure
+//
+// This file is distributed under the University of Illinois Open Source
+// License. See LICENSE.TXT for details.
+//
+//===------------------------------------------------------------------------===//
+//
+// This file implements StructFieldAccessInfo that is used in StructFieldCacheAnalysis pass.
+//
+//===------------------------------------------------------------------------===//
+
+#include "StructFieldCacheAnalysisImpl.h"
+#include "llvm/IR/Instructions.h"
+#include "llvm/IR/Operator.h"
+
+using namespace llvm;
+
+#define DEBUG_TYPE "struct-analysis"
+#define DEBUG_TYPE_IR "struct-analysis-IR"
+#define DEBUG_TYPE_STATS "struct-analysis-detailed-stats"
+
+// Functions for StructFieldAccessInfo
+void StructFieldAccessInfo::addFieldAccessNum(const Instruction *I,
+                                              FieldNumType FieldNum) {
+  assert(I->getOpcode() == Instruction::Load ||
+         I->getOpcode() == Instruction::Store); // Only loads and stores
+  assert(LoadStoreFieldAccessMap.find(I) == LoadStoreFieldAccessMap.end());
+  LoadStoreFieldAccessMap[I] = FieldNum;
+}
+
+void StructFieldAccessInfo::addFieldAccessNum(const Instruction* I, const Function* F, unsigned Arg, FieldNumType FieldNum)
+{
+  assert(I->getOpcode() == Instruction::Call || I->getOpcode() == Instruction::Invoke); // Only calls and invokes
+  if (F->isDeclaration())
+    // Give up if the function only has declaration
+    return;
+  if (CallInstFieldAccessMap.find(I) == CallInstFieldAccessMap.end()){
+    CallInstFieldAccessMap[I] = new FunctionCallInfo(F, Arg, FieldNum);
+  }
+  else{
+    auto* CallSite = CallInstFieldAccessMap[I];
+    assert(CallSite->FunctionDeclaration == F);
+    CallSite->insertCallInfo(Arg, FieldNum);
+  }
+}
+
+Optional<FieldNumType>
+StructFieldAccessInfo::getAccessFieldNum(const Instruction *I) const {
+  Optional<FieldNumType> ret;
+  auto it = LoadStoreFieldAccessMap.find(I);
+  if (it != LoadStoreFieldAccessMap.end()) {
+    ret = it->second;
+  }
+  return ret;
+}
+
+FieldNumType
+StructFieldAccessInfo::calculateFieldNumFromGEP(const User *U) const {
+  DEBUG_WITH_TYPE(DEBUG_TYPE_IR, dbgs() << "Calculating field number from GEP: "
+                                        << *U << "\n");
+  // Operand 0 should be a pointer to the struct
+  assert(isa<GetElementPtrInst>(U) || isa<GEPOperator>(U));
+  auto *Op = U->getOperand(0);
+  // Make sure Operand 0 is a struct type and matches the current struct type of
+  // StructFieldAccessInfo
+  assert(Op->getType()->getPointerElementType()->isStructTy() &&
+         Op->getType()->getPointerElementType() == StructureType);
+  if (U->getNumOperands() < 3) // GEP to calculate struct field needs at least 2
+                               // indices (operand 1 and 2)
+    return 0;
+  // Operand 1 should be first index to the struct, usually 0; if not 0, it's
+  // like goto an element of an array of structs
+  Op = U->getOperand(1);
+  // TODO: ignore this index for now because it's the same for an array of
+  // structs
+  assert(Op->getType()->isIntegerTy());
+  // Operand 2 should be the index to the field, and be a constant
+  Op = U->getOperand(2);
+  assert(isa<Constant>(Op));
+  auto *Index = dyn_cast<Constant>(Op);
+  auto Offset = (FieldNumType)Index->getUniqueInteger().getZExtValue();
+  assert(Offset < NumElements);
+  // TODO: ignore indices after this one. If there's indices, the field has to
+  // be an array or struct
+  return Offset + 1; // return field number starting from 1
+}
+
+void StructFieldAccessInfo::addFieldAccessFromGEP(const User *U) {
+  DEBUG_WITH_TYPE(DEBUG_TYPE_IR,
+                  dbgs() << "Analyze all users of GEP: " << *U << "\n");
+  assert(isa<GetElementPtrInst>(U) || isa<GEPOperator>(U));
+  auto FieldLoc = calculateFieldNumFromGEP(U);
+  if (FieldLoc == 0)
+    return;
+  for (auto *User : U->users()) {
+    DEBUG_WITH_TYPE(DEBUG_TYPE_IR,
+                    dbgs() << "Check user of " << *U << ": " << *User << "\n");
+    if (isa<Instruction>(User)) {
+      auto *Inst = dyn_cast<Instruction>(User);
+      if (Inst->getOpcode() == Instruction::Load)
+        addFieldAccessNum(Inst, FieldLoc);
+      else if (Inst->getOpcode() == Instruction::Store) {
+        if (U == Inst->getOperand(1))
+          addFieldAccessNum(Inst, FieldLoc);
+      }
+      else{
+        if (Inst->getOpcode() == Instruction::Call){
+          auto* Call = dyn_cast<CallInst>(Inst);
+          auto* Func = Call->getCalledFunction();
+          if (Func && Func->arg_size() == Call->getNumArgOperands()){
+            for (unsigned i = 0; i < Call->getNumArgOperands(); i++){
+              if (Call->getArgOperand(i) == U)
+                addFieldAccessNum(Inst, Func, i, FieldLoc);
+            }
+          }
+          else{
+            // Bail out if Gep is used in an indirect function or a function that has undetermined args
+            addStats(StructFieldAccessManager::DebugStats::DS_GepPassedIntoIndirectFunc);
+          }
+        }
+        else if (Inst->getOpcode() == Instruction::Invoke){
+          auto* Call = dyn_cast<InvokeInst>(Inst);
+          auto* Func = Call->getCalledFunction();
+          if (Func){
+            for (unsigned i = 0; i < Call->getNumArgOperands(); i++){
+              if (Call->getArgOperand(i) == U)
+                addFieldAccessNum(Inst, Func, i, FieldLoc);
+            }
+          }
+          else{
+            addStats(StructFieldAccessManager::DebugStats::DS_GepPassedIntoIndirectFunc);
+          }
+        }
+        else if (Inst->getOpcode() == Instruction::BitCast){
+          addStats(StructFieldAccessManager::DebugStats::DS_GepPassedIntoBitcast);
+        }
+        else{
+          // TODO: Collect stats of this kind of access and add analysis later
+          addStats(StructFieldAccessManager::DebugStats::DS_GepUnknownUse,
+                   Inst->getOpcode());
+        }
+      }
+    } else if (isa<Operator>(User)) {
+      auto *Inst = dyn_cast<Operator>(User);
+      if (Inst->getOpcode() == Instruction::BitCast) {
+        addStats(StructFieldAccessManager::DebugStats::DS_GepPassedIntoBitcast);
+      } else {
+        // TODO: Collect stats of this kind of access and add analysis later
+        addStats(StructFieldAccessManager::DebugStats::DS_GepUnknownUse,
+                 Inst->getOpcode());
+      }
+    } else {
+      addStats(StructFieldAccessManager::DebugStats::
+                   DS_UserNotInstructionNorOperator);
+    }
+  }
+}
+
+void StructFieldAccessInfo::analyzeUsersOfStructValue(const Value *V) {
+  assert(V->getType()->isPointerTy() &&
+         V->getType()->getPointerElementType()->isStructTy());
+  for (auto *U : V->users()) {
+    DEBUG_WITH_TYPE(DEBUG_TYPE_IR,
+                    dbgs() << "Analyzing user of " << *V << ": " << *U << "\n");
+    if (isa<Instruction>(U)) {
+      auto *Inst = dyn_cast<Instruction>(U);
+      if (!isa<GetElementPtrInst>(Inst)) {
+        // Only support access struct through GEP for now
+        if (Inst->getOpcode() == Instruction::Call){
+          DEBUG_WITH_TYPE(DEBUG_TYPE_IR, dbgs() << "User is a call instruction\n");
+          assert(Inst && isa<CallInst>(Inst));
+          auto* F = dyn_cast<CallInst>(Inst)->getCalledFunction();
+          if (!F || F->isDeclaration()){
+            // If a struct is passed to an indirect or a function not declared in the program, we can't analyze it...
+            Eligiblity = false;
+          }
+        }
+        else if (Inst->getOpcode() == Instruction::Invoke){
+          DEBUG_WITH_TYPE(DEBUG_TYPE_IR, dbgs() << "User is an invoke instruction\n");
+          assert(Inst && isa<CallInst>(Inst));
+          auto* F = dyn_cast<InvokeInst>(Inst)->getCalledFunction();
+          if (!F || F->isDeclaration()){
+            // If a struct is passed to an indirect or a function not declared in the program, we can't analyze it...
+            Eligiblity = false;
+          }
+        }
+        continue;
+      }
+      addFieldAccessFromGEP(Inst);
+    } else if (isa<Operator>(U)) {
+      auto *Inst = dyn_cast<Operator>(U);
+      if (!isa<GEPOperator>(Inst)) {
+        // Only support access struct through GEP for now
+        continue;
+      }
+      addFieldAccessFromGEP(Inst);
+    } else {
+      addStats(StructFieldAccessManager::DebugStats::
+                   DS_UserNotInstructionNorOperator);
+    }
+  }
+}
+
+void StructFieldAccessInfo::analyzeUsersOfStructPointerValue(const Value *V) {
+  // Analyze users of value defined as struct*
+  assert(V->getType()->isPointerTy() &&
+         V->getType()->getPointerElementType()->isPointerTy() &&
+         V->getType()
+             ->getPointerElementType()
+             ->getPointerElementType()
+             ->isStructTy());
+  for (auto *U : V->users()) {
+    DEBUG_WITH_TYPE(DEBUG_TYPE_IR,
+                    dbgs() << "Analyzing user of " << *V << ": " << *U << "\n");
+    if (isa<Instruction>(U)) {
+      auto *Inst = dyn_cast<Instruction>(U);
+      if (Inst->getOpcode() == Instruction::Load) {
+        analyzeUsersOfStructValue(Inst);
+      } else if (Inst->getOpcode() == Instruction::GetElementPtr) {
+        addStats(StructFieldAccessManager::DebugStats::DS_GepUsedOnStructPtr);
+      } else {
+        addStats(
+            StructFieldAccessManager::DebugStats::DS_UnknownUsesOnStructPtr);
+      }
+    } else if (isa<Operator>(U)) {
+      auto *Inst = dyn_cast<Operator>(U);
+      if (Inst->getOpcode() == Instruction::Load) {
+        analyzeUsersOfStructValue(Inst);
+      } else if (Inst->getOpcode() == Instruction::GetElementPtr) {
+        addStats(StructFieldAccessManager::DebugStats::DS_GepUsedOnStructPtr);
+      } else {
+        addStats(
+            StructFieldAccessManager::DebugStats::DS_UnknownUsesOnStructPtr);
+      }
+    } else {
+      addStats(StructFieldAccessManager::DebugStats::
+                   DS_UserNotInstructionNorOperator);
+    }
+  }
+}
+
+void StructFieldAccessInfo::summarizeFunctionCalls()
+{
+  DEBUG_WITH_TYPE(DEBUG_TYPE_IR, dbgs() << "Summarizes call/invokes instructions into function declarations\n");
+  for (auto& it : CallInstFieldAccessMap){
+    auto* CallSiteInfo = it.second;
+    assert(CallSiteInfo);
+    auto* F = CallSiteInfo->FunctionDeclaration;
+    assert(F);
+    if (FunctionAccessMap.find(F) == FunctionAccessMap.end()){
+      FunctionAccessMap[F] = new FunctionAccessPattern(&CallSiteInfo->Arguments);
+    }
+    else{
+      FunctionAccessMap[F]->insertCallInfo(&CallSiteInfo->Arguments);
+    }
+  }
+  for (auto& it : FunctionAccessMap){
+    DEBUG_WITH_TYPE(DEBUG_TYPE_IR, dbgs() << "Function " << it.first->getName() << " is called with fields as argument:\n");
+    for (auto* Args : it.second->CallSites){
+      for (unsigned i = 0; i < Args->size(); i++){
+        if ((*Args)[i] > 0)
+          DEBUG_WITH_TYPE(DEBUG_TYPE_IR, dbgs() << "Arg " << i << " is called as field " << (*Args)[i] << "\n");
+      }
+    }
+  }
+
+}
+
+ExecutionCountType StructFieldAccessInfo::calculateTotalHotness() const
+{
+  ExecutionCountType Hotness = 0;
+  for (auto &it : LoadStoreFieldAccessMap){
+    auto Count = getExecutionCount(it.first);
+    if (Count.hasValue())
+      Hotness += Count.getValue();
+  }
+  for (auto &it : CallInstFieldAccessMap){
+    auto Count = getExecutionCount(it.first);
+    if (Count.hasValue())
+      Hotness += Count.getValue();
+  }
+  return Hotness;
+}
+
+void StructFieldAccessInfo::debugPrintAllStructAccesses(raw_ostream &OS) {
+  for (auto &it : LoadStoreFieldAccessMap) {
+    OS << "\tInstruction [" << *it.first << "] accesses field number ["
+       << it.second << "]\n";
+  }
+  for (auto &it : CallInstFieldAccessMap){
+    OS << "\tInstruction [" << *it.first << "] calls function " << it.second->FunctionDeclaration->getName() << " access fields\n";
+  }
+}
+
+// Functions for StructHotnessAnalyzer
+void StructHotnessAnalyzer::addStruct(const StructFieldAccessInfo* SI)
+{
+  auto Hotness = SI->calculateTotalHotness();
+  if (Hotness > MaxHotness)
+    MaxHotness = Hotness;
+  TotalHotnessToSort.push_back(Hotness);
+}
+
+void StructHotnessAnalyzer::summarize()
+{
+  // TODO: vary the number of buckets
+  Histogram.resize(HistogramSize);
+  MaxHotness += 1; // To avoid the largest one out of bound
+  for (auto Hotness : TotalHotnessToSort){
+    auto Index = Hotness * HistogramSize / MaxHotness;
+    Histogram[Index]++;
+  }
+  dbgs() << "Distribution of struct hotness: \n";
+  for (unsigned i = 0; i < Histogram.size(); i++)
+    dbgs() << "Hotness >=" << MaxHotness * i << ": " << Histogram[i] << "\n";
+}
